requirements.txt
python-telegram-bot
cloudmersive-ocr-api-client